move tasks actions to slice
